chore(automation): reorganize automation code, split into separate files by commands
